# Replace s2n-bignum's Montgomery reduction with its latest, faster version